% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetsar.R
\name{hetsar}
\alias{hetsar}
\title{Heterogeneous SAR estimation}
\usage{
hetsar(
  data,
  W,
  indices,
  dependent,
  explanatory = NULL,
  nocons = FALSE,
  time_lags = NULL,
  space_lags = NULL,
  time_space_lags = NULL,
  verbose = "silent",
  bounds = NULL
)
}
\arguments{
\item{data}{A \code{data.frame}, in tidy format. The panel \emph{must} be balanced.}

\item{W}{Spatial weight \code{matrix}. Rows and columns should be ordered such that colnames(W)==rowanems(W)==unique(df$id).}

\item{indices}{Character vector of length 2 with the names of the columns in \code{data} containing the unit-identifier (\code{indices[[1]]}) and the time-identifier (\code{indices[[2]]}).}

\item{dependent}{Name of the dependent variable.}

\item{explanatory}{Character vector with the names of the explanatory variables, excluding: (a) the spatial lag of the dependent variable \code{Wy}, included by default; (b) the intercept; (c) space/time-lags. See Details.}

\item{nocons}{Logical: Shall the constant term (i.e. the intercept) be excluded?}

\item{time_lags}{Character vector with the names of the time-lagged variables.}

\item{space_lags}{Character vector with the names of the space-lagged variables.}

\item{time_space_lags}{Character vector with the names of the time- and space-lagged variables.}

\item{verbose}{Character vector. Verbosely print some extra information relative to the model specification (\code{verbose="model"}), to the bounds (\code{verbose="bounds"}), to the optimisation procedure (\code{verbose="optimisation"}), or to all the above (\verb{verbose="all"). Default is silent (}verbose="silent"`).}

\item{bounds}{(Advanced) Named-\code{list} specifying the parameter space, see "Bounds".}
}
\value{
TODO An object (\code{list}) of class \code{hetsar} with the following elements:
\item{coefficients}{A matrix of estimated coefficients. !! COLNAMES}
\item{Std.err.sandwich}{A matrix with standard errors computed using the sandwich formula.}
\item{Std.err.standard}{A matrix with standard errors computed using the standard formula.}
\item{N}{The number of spatial units.}
\item{T}{The number of time periods.}
\item{NT}{The number of observations.}
}
\description{
Estimation of heterogeneous spatial autoregressive models
}
\section{Model specification}{

TODO
}

\section{Bounds}{

TODO
}

\examples{

data(hetsarDataDemo)
df_data <- hetsarDataDemo[["data"]]
m_C <- hetsarDataDemo[["network_matrix"]]
m_W <- m_C / rowSums(m_C) # row normalise

out <- hetsar(
  data = df_data, 
  W = m_W,
  indices = c("id", "time"),
  dependent = "y",
  explanatory = "x",
  space_lags = "x")

summary(out)
summary(out, MG = TRUE)

}
\references{
M. Aquaro, N. Bailey and M. H. Pesaran (2021). "Estimation and inference for spatial models with heterogeneous coefficients: An application to US house prices". Journal of Applied Econometrics 36(1): 18-44. doi:https://doi.org/10.1002/jae.2792
}
\seealso{
TODO See also \code{\link[hetsar]{summary.hetsar}} to see the results.
}
\author{
Michele Aquaro
}
