% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hetsarDataDemo}
\alias{hetsarDataDemo}
\title{Demo data for the hetsar package}
\format{
A list with 2 items:
\describe{
\item{data}{A data-frame of 5025 rows and 4 columns of simulated data:}
\describe{
\item{time}{Time identifier}
\item{id}{Unit identifier}
\item{y}{Dependent variable}
\item{x}{Exogenous variable}
}
\item{network_matrix}{A simulated network matrix with 25 rows and 25 columns}
}
}
\usage{
hetsarDataDemo
}
\description{
Simulated data to illustrate how to use the hetsar package.
}
\keyword{datasets}
